portfolio: Add a tags system that uses js to change portfolio
This is a lot of javascript that currently implements a successful
tagging system to only display some experience elements in categories.
This is an initial commit because the JS has some inefficiency and
also needs cleaning up, comments, and stylistic changes.
Also, in the future, add some js to make the tags at the top
of the page displayed dynamically based on the existing tags in the
posts.